Refactor f classes to fix declaration hell
