+++ include/dol/base/f_list_mg.hpp
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <types.h>
+#include <dol/profile/profile.hpp>
+#include <dol/cLib/c_list.hpp>
+
+class fLiMgBa_c : public cListMg_c {
+public:
+    fLiMgBa_c() {
+        mpFirst = nullptr;
+        mpLast = nullptr;
+    }
+};
