+++ include/dol/base/f_list_nd.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <types.h>
+#include <dol/cLib/c_list.hpp>
+
+class fLiNdBa_c : public cListNd_c {};
