+++ include/dol/base/f_list_nd_prio.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <types.h>
+#include <dol/profile/profile.hpp>
+#include <dol/base/f_list_nd.hpp>
+
+// Unofficial name
+class fLiNdPrio_c : fLiNdBa_c {
+public:
+    short mPriority;
+};
