+++ source/dol/base/f_list.cpp
@@ -1,7 +1,7 @@
-#include <dol/base/f_list.hpp>
 #include <types.h>
+#include <dol/base/f_list_mg.hpp>
 
 fLiMgBa_c::fLiMgBa_c() {
-    this->mpTail = nullptr;
-    this->mpHead = nullptr;
+    this->mpFirst = nullptr;
+    this->mpLast = nullptr;
 }
